Refactored ContactControllerTest, ContactServiceTest.
